+++ src/app/helpers/mapbox-layer-helpers.ts
@@ -2,22 +2,22 @@ import * as mapboxgl from "mapbox-gl";
 
 import OJP_Legacy from '../config/ojp-legacy';
 
-import { MapLegTypeColor, MapLegTypes, MapLegLineTypeColor } from '../config/map-colors';
+import { MapLegLineTypeColor } from '../config/map-colors';
 
 import { TripLegDrawType, TripLegLineType, TripLegPropertiesEnum } from '../shared/types/map-geometry-types';
 
 export class MapboxLayerHelpers {
   public static FilterBeelines(): mapboxgl.ExpressionSpecification {
-    const drawType: TripLegDrawType = 'Beeline'
-    return this.FilterByDrawType(drawType)
+    const drawType: TripLegDrawType = 'Beeline';
+    return this.FilterByDrawType(drawType);
   }
 
   public static FilterWalkingLegs(): mapboxgl.ExpressionSpecification {
     const filterExpression: mapboxgl.ExpressionSpecification = [
       "all",
       this.FilterByDrawType('LegLine'),
-      this.FilterByLineType('Walk'),
-    ]
+      this.FilterByLineTypes(['Guidance', 'Walk']),
+    ];
 
     return filterExpression
   }
@@ -58,37 +58,31 @@ export class MapboxLayerHelpers {
     return filterExpression
   }
 
-  private static FilterByLineType(lineType: TripLegLineType): mapboxgl.ExpressionSpecification {
-    const filterExpression: mapboxgl.ExpressionSpecification = [
-      "==", ["get", TripLegPropertiesEnum.LineType], lineType
-    ]
-    return filterExpression
-  }
-
-  public static FilterTimedLegTracks(): mapboxgl.ExpressionSpecification {
+  public static FilterLegTracks(excludeLineTypes: TripLegLineType[]): mapboxgl.ExpressionSpecification {
     const filterExpression: mapboxgl.ExpressionSpecification = [
       "all",
-      // TODO - exclude Walk
       this.FilterByDrawType('LegLine'),
-    ]
+    ];
+
+    if (excludeLineTypes.length > 0) {
+      // ["!", ["in", ["get", "prop2"], ["literal", ["value2", "value3"]]]]
+      const excludeLineTypesExpression: mapboxgl.ExpressionSpecification = [
+        "!",
+        this.FilterByLineTypes(['Guidance', 'Walk']),
+      ];
+      filterExpression.push(excludeLineTypesExpression);
+    }
 
     return filterExpression
   }
 
-  public static ColorCaseByLegType(): mapboxgl.ExpressionSpecification {
-    const caseExpression: mapboxgl.ExpressionSpecification = ["case"]
+  private static FilterByLineTypes(lineTypes: TripLegLineType[]): mapboxgl.ExpressionSpecification {
+    const lineTypesExpression: mapboxgl.ExpressionSpecification = [
+      "in",
+      ["get", TripLegPropertiesEnum.LineType],
+      ["literal", lineTypes],
+    ];
 
-    MapLegTypes.forEach(legType => {
-      const caseOptionCondition = ["==", ["get", TripLegPropertiesEnum.LegType], legType]
-      caseExpression.push(caseOptionCondition)
-
-      const colorCode = MapLegTypeColor[legType]
-      caseExpression.push(colorCode)
-    });
-
-    // Default is Pink
-    caseExpression.push('#FF1493')
-
-    return caseExpression
+    return lineTypesExpression;
   }
 }